implement config.Reporter with envconfig
The current implementation of config.Reporter is one we'd to deprecate. These
changes show one way to start migrating away. The use of reflection is
effective, but probably not ideal.

It might be better to look at the tags on the fields themselves, as they can
be entirely different from the property names.