implement config.Reporter with envconfig (no reflection)
The current implementation of config.Reporter is one we'd to deprecate. These
changes show one way to start migrating away. Here we use a manual match of
the key with a parameter. It is far from ideal, but if we assume that
client.Config doesn't change often, perhaps it would work fine? It might need
to be adjusted for scopes or something, but the concept is there.